Added support to send UDP Broadcasts with a payload to allow for sending responses, or could be allowed to provide information from the game. Added new configuration for the UDP Broadcast port Changed Default ports to 12333 and 12334 which are listed as unassigned by IANA Reworked and added custom log handler
Decided that the rename of the package was best solution, since doing some changes to building.
Also since this hasn't had a public mod.io release this should be fine as this plugin has gone virtually unused.
It will break any plugins if used as a dependency(unlikely)
Version bump to 2.0.0 due to the breaking nature of package change